requirements.txt
python-telegram-bot[job-queue]==22.1
requests==2.32.3
pydantic==2.8.2
pydantic-settings==2.4.0
sentry-sdk==2.8.0
tenacity==9.0.0
python-json-logger==2.0.7
psycopg[binary,pool]==3.2.1
